+++ src/AngleSharp.Css/Dom/Internal/Rules/CssPageRule.cs
@@ -34,7 +34,7 @@ namespace AngleSharp.Css.Dom
         public String SelectorText
         {
             get => _selector?.Text;
-            set { _selector = ParseSelector(value); }
+            set { _selector = ParseSelector(value); MarkChanged(); }
         }
 
         public ISelector Selector => _selector;