+++ src/AngleSharp.Css/Dom/Internal/Rules/CssSupportsRule.cs
@@ -56,6 +56,7 @@ namespace AngleSharp.Css.Dom
             }
 
             _condition = condition;
+            MarkChanged();
             return true;
         }